requirements.txt
pandas
google-cloud-bigquery
streamlit
google-cloud-discoveryengine
db-dtypes